backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pypdf==4.2.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy>=1.24.0
openai>=1.0.0
